Code/aiplot-eval-master/pyproject.toml
[project]
name = "aiplot-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "instructor>=1.7.9",
    "openai>=1.75.0",
    "pyyaml>=6.0.2",
    "streamlit>=1.44.1",
]
